lib/uklcpu: Delete no longer necessary `core.h`
Initially, `core.h` was added as an ugly temporary workaround to
solve the cyclical dependency between platform code implementing
LCPU PAL operations while still needing access to LCPU definitions.
The aforementioned header would define the very basic/barebones
definitions that had no dependencies so that platform code could
include it.

Now that there is no such code referencing LCPU definitions, remove
this ugly header!

Checkpatch-Ignore: LONG_LINE_COMMENT
Checkpatch-Ignore: VOLATILE
