+++ lib/uklcpu/include/uk/lcpu.h
@@ -1,5 +1,6 @@
 /* SPDX-License-Identifier: BSD-3-Clause */
 /*
+ * Copyright (c) 2018, NEC Europe Ltd., NEC Corporation. All rights reserved.
  * Copyright (c) 2022, Karlsruhe Institute of Technology (KIT)
  *                     All rights reserved.
  * Copyright (c) 2022, University Politehnica of Bucharest.
@@ -12,9 +13,9 @@
 #ifndef __UK_LCPU_H__
 #define __UK_LCPU_H__
 
+#include <uk/arch.h>
 #include <uk/arch/types.h>
 #include <uk/pcpuvar.h>
-#include <uk/lcpu/core.h>
 #include <uk/plat/pal/lcpu.h>
 #include <uk/pal/lcpu.h>
 #include <uk/lcpu/ectx.h>
@@ -24,17 +25,174 @@
 #include <uk/lcpu/sysctx.h>
 #include <uk/lcpu/pm.h>
 
-#if !__ASSEMBLY__
-#include <uk/essentials.h>
-
 #ifdef __cplusplus
 extern "C" {
 #endif
+/*
+ * UK_LCPU Startup Arguments
+ */
+#define UK_LCPU_SARGS_ENTRY_OFFSET	0x00
+#define UK_LCPU_SARGS_STACKP_OFFSET					\
+	(UK_LCPU_SARGS_ENTRY_OFFSET + 0x08)
+
+#define UK_LCPU_SARGS_SIZE		0x10
+
+#if !__ASSEMBLY__
+struct uk_lcpu_sargs {
+	__uptr entry;
+	__uptr stackp;
+};
+
+UK_CTASSERT(__offsetof(struct uk_lcpu_sargs, entry)  ==
+	    UK_LCPU_SARGS_ENTRY_OFFSET);
+UK_CTASSERT(__offsetof(struct uk_lcpu_sargs, stackp) ==
+	    UK_LCPU_SARGS_STACKP_OFFSET);
+
+UK_CTASSERT(sizeof(struct uk_lcpu_sargs) == UK_LCPU_SARGS_SIZE);
+#endif /* !__ASSEMBLY__ */
+
+#if CONFIG_HAVE_SMP
+#define UK_LCPU_FUNC_SIZE		0x10
+
+#if !__ASSEMBLY__
+struct uk_lcpu_regs;
+
+struct uk_lcpu_func {
+	/**
+	 * Function to execute.
+	 *
+	 * @param regs pointer to a snapshot of the current CPU register state.
+	 *    Changes to the registers are applied after the RUN IRQ handler
+	 *    returns. The parameter might be NULL if the platform does not
+	 *    support supplying a register snapshot.
+	 * @param arg user-supplied argument
+	 */
+	void (*fn)(struct uk_lcpu_regs *regs, void *arg);
+
+	/* Optional user-supplied argument. */
+	void *user;
+};
+
+UK_CTASSERT(sizeof(struct uk_lcpu_func) == UK_LCPU_FUNC_SIZE);
+#endif /* !__ASSEMBLY__ */
+#endif /* CONFIG_HAVE_SMP */
+
+/*
+ * Logical CPU (LCPU) Structure
+ */
+#define UK_LCPU_STATE_OFFSET		0x00
+#define UK_LCPU_ENTRY_OFFSET		(UK_LCPU_STATE_OFFSET + 0x08)
+#define UK_LCPU_STACKP_OFFSET		(UK_LCPU_ENTRY_OFFSET + 0x08)
+#define UK_LCPU_ERR_OFFSET		(UK_LCPU_ENTRY_OFFSET + 0x0)
+
+#define UK_LCPU_MEMBERS_SIZE						\
+	(UK_LCPU_STACKP_OFFSET  + 8)
+#define UK_LCPU_SIZE							\
+	(ALIGN_UP(UK_LCPU_MEMBERS_SIZE, UK_ARCH_CACHE_LINE_SIZE))
+
+/**
+ * UK_LCPU States
+ * The following state transitions are safe to execute.
+ *
+ *                         lcpu_init
+ *                   ┌───────────────────┐lcpu_run
+ *        lcpu_start │          ┌──────┐ │ ┌─────┐   ┌────
+ * ┌─────────┐   ┌───┴──┐   ┌───┴──┐ ┌─▼─▼─┴─┐ ┌─▼───┴─┐
+ * │ OFFLINE ├──►│ INIT │   │ IDLE │ │ BUSY0 │ │ BUSY1 │
+ * └─────────┘   └───┬──┘   └─┬─▲──┘ └─┬─┬─▲─┘ └─┬─┬─▲─┘
+ *                   │        │ └──────┘ │ └─────┘ │ └────
+ * ┌────────┐        │        │          │ RUN_IRQ │
+ * │ HALTED │◄───────┴────────┴──────────┴─────────┴──────
+ * └────────┘        lcpu_halt (ONLY ALLOWED FOR THIS CPU)
+ */
+#define UK_LCPU_STATE_HALTED		-1 /* CPU stopped, needs reset */
+#define UK_LCPU_STATE_OFFLINE		0 /* CPU not started */
+#define UK_LCPU_STATE_INIT		1 /* CPU started, init not finished */
+#define UK_LCPU_STATE_IDLE		2 /* CPU is idle */
+#define UK_LCPU_STATE_BUSY0		3 /* >= CPU is busy */
+
+#if !__ASSEMBLY__
+struct __align(UK_ARCH_CACHE_LINE_SIZE) uk_lcpu {
+	/* Current CPU state (UK_LCPU_STATE_*).
+	 * Typically accessed with atomic instructions - must be 8-byte aligned
+	 */
+	volatile int state __align(8);
+
+	union {
+		/* Startup arguments
+		 * Only valid in UK_LCPU_STATE_INIT
+		 */
+		struct uk_lcpu_sargs s_args;
+
+		/* Remote function to execute
+		 * Only valid in UK_LCPU_STATE_IDLE and busy states
+		 */
+#if CONFIG_HAVE_SMP
+		struct uk_lcpu_func fn;
+#endif /* CONFIG_HAVE_SMP */
+
+		/* Error code indicating the halt reason
+		 * Only valid in UK_LCPU_STATE_HALTED
+		 */
+		int error_code;
+	};
+};
+
+UK_CTASSERT(__offsetof(struct uk_lcpu, state) == UK_LCPU_STATE_OFFSET);
+UK_CTASSERT(__offsetof(struct uk_lcpu, s_args.entry) == UK_LCPU_ENTRY_OFFSET);
+UK_CTASSERT(__offsetof(struct uk_lcpu, s_args.stackp) == UK_LCPU_STACKP_OFFSET);
+UK_CTASSERT(__offsetof(struct uk_lcpu, error_code) == UK_LCPU_ERR_OFFSET);
+
+UK_CTASSERT(sizeof(struct uk_lcpu) == UK_LCPU_SIZE);
+UK_CTASSERT(UK_LCPU_MEMBERS_SIZE <= UK_LCPU_SIZE);
+#endif /* !__ASSEMBLY__ */
+
+#if !__ASSEMBLY__
+#include <uk/arch/time.h>
+#include <uk/essentials.h>
+
+/**
+ * Return the UK_LCPU structure for the CPU executing this function
+ */
+struct uk_lcpu *uk_lcpu_get_current(void);
+
+extern __uk_pcpuvar struct uk_lcpu uk_lcpus;
+
+/**
+ * Return the UK_LCPU structure for the bootstraping CPU
+ */
+static inline struct uk_lcpu *uk_lcpu_get_bsp(void)
+{
+	/* The BSP is always index 0 */
+	return &uk_pcpuvar_lval(0, uk_lcpus);
+}
+
+/**
+ * Return 1 if the supplied UK_LCPU is the boottrap processor, 0 otherwise
+ */
+static inline __bool uk_lcpu_is_bsp(struct uk_lcpu *lcpu)
+{
+	return (lcpu == uk_lcpu_get_bsp());
+}
+
+/**
+ * Return 1 if the executed on the bootstrap processor, 0 otherwise
+ */
+static inline __bool uk_lcpu_current_is_bsp(void)
+{
+	return uk_lcpu_is_bsp(uk_lcpu_get_current());
+}
+
+#if CONFIG_HAVE_SMP
+/* The IRQ vectors passed to lcpu_mp_init */
+extern const unsigned long * const uk_lcpu_run_irqv;
+extern const unsigned long * const uk_lcpu_wakeup_irqv;
+#endif /* CONFIG_HAVE_SMP */
 
 /**
  * Return 1 if the given UK_LCPU is online, 0 otherwise
  */
-static inline int uk_lcpu_state_is_online(int state)
+static inline __bool uk_lcpu_state_is_online(int state)
 {
 	return (state >= UK_LCPU_STATE_IDLE);
 }
@@ -43,7 +201,7 @@ static inline int uk_lcpu_state_is_online(int state)
  * Return 1 if the given UK_LCPU is busy, 0 otherwise.
  * NOTE: The negation (i.e., the UK_LCPU is idle) does not have be true!
  */
-static inline int uk_lcpu_state_is_busy(int state)
+static inline __bool uk_lcpu_state_is_busy(int state)
 {
 	return (state >= UK_LCPU_STATE_BUSY0);
 }
@@ -73,9 +231,9 @@ int uk_lcpu_init(struct uk_lcpu *this_lcpu);
 int uk_lcpu_mp_init(unsigned long run_irq, unsigned long wakeup_irq, void *arg);
 
 /**
- * Default entry function for secondary logical CPUs. Will call lcpu_init() and
- * If the logical CPU's startup arguments supply an entry function, the
- * original stack pointer will be restored and execution continues in the
+ * Default entry function for secondary logical CPUs. Calls uk_lcpu_init() and
+ * if the logical CPU's startup arguments supply an entry function, the
+ * original stack pointer is restored and execution continues in the
  * supplied entry function with interrupts still disabled. Otherwise, interrupts
  * are enabled and the CPU enters a low-power state to wait for interrupts and
  * calls of uk_lcpu_run() that are destined for this CPU.
@@ -171,8 +329,8 @@ struct uk_lcpu *uk_lcpu_get_current(void);
  */
 __isr __u64 uk_lcpu_get_current_idx_in_except(void);
 
+#endif /* !__ASSEMBLY__ */
 #ifdef __cplusplus
 }
 #endif
-#endif /* !__ASSEMBLY__ */
 #endif /* __UK_LCPU_H__*/
